pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["LilDrugHill <smookigrow@icloud.com>"]
packages = [{ include = 'task_manager' }]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
gunicorn = "^20.1.0"
django-on-heroku = "^1.1.2"
django-bootstrap4 = "^22.2"
python-dotenv = "^0.21.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
heroku = "^0.1.4"
rollbar = "^0.16.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
